# apps/backend/openapi-quotes.yaml
openapi: 3.0.3
info:
  title: Pivotal Flow Quotes API
  description: |
    API for managing quotes in the Pivotal Flow system.
    
    ## Features
    - Create, read, update, and delete quotes
    - Quote status transitions (draft → pending → approved → sent → accepted)
    - Automatic calculations (subtotal, tax, discounts, total)
    - Multi-tenant support with organization isolation
    - Audit logging for all quote changes
    - Quote number generation with organization prefixes
    
    ## Authentication
    All endpoints require authentication via JWT token in the Authorization header.
    The token must include organization and user context for multi-tenant access.
    
    ## Multi-tenancy
    All quotes are scoped to the user's organization. Users can only access quotes
    within their own organization.
  version: 1.0.0
  contact:
    name: Pivotal Flow Team
    email: support@pivotalflow.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.pivotalflow.com
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token containing user and organization context.
        Token must include:
        - `sub`: User ID
        - `org`: Organization ID
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
  schemas:
    QuoteStatus:
      type: string
      enum: [draft, pending, approved, sent, accepted, rejected, cancelled]
      description: Current status of the quote
      example: "draft"
    QuoteType:
      type: string
      enum: [project, service, product, recurring, one_time]
      description: Type of quote
      example: "project"
    LineItemType:
      type: string
      enum: [service, product, material, travel, expense, discount, tax]
      description: Type of line item
      example: "service"
    DiscountType:
      type: string
      enum: [percentage, fixed_amount, per_unit]
      description: Type of discount
      example: "percentage"
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: string
          description: Amount as a decimal string for precision
          example: "100.00"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: ISO 4217 currency code
          example: "NZD"
    LineItem:
      type: object
      required:
        - lineNumber
        - description
        - quantity
        - unitPrice
      properties:
        lineNumber:
          type: integer
          minimum: 1
          description: Sequential line number
          example: 1
        type:
          $ref: '#/components/schemas/LineItemType'
        description:
          type: string
          minLength: 1
          maxLength: 1000
          description: Item description
          example: "Web Development Services"
        quantity:
          type: number
          minimum: 0
          description: Quantity of items
          example: 40
        unitPrice:
          $ref: '#/components/schemas/Money'
        unitCost:
          $ref: '#/components/schemas/Money'
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          description: Tax rate as decimal (0.15 = 15%)"
          example: 0.15
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Discount value
          example: 0
        serviceCategoryId:
          type: string
          format: uuid
          description: Service category ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        rateCardId:
          type: string
          format: uuid
          description: Rate card ID for pricing
          example: "550e8400-e29b-41d4-a716-446655440001"
        metadata:
          type: object
          description: Additional metadata
          example: { "projectPhase": "development", "priority": "high" }
    Quote:
      type: object
      required:
        - id
        - quoteNumber
        - customerId
        - title
        - status
        - type
        - validFrom
        - validUntil
        - currency
        - subtotal
        - taxAmount
        - totalAmount
        - createdBy
        - organizationId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique quote identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        quoteNumber:
          type: string
          description: Human-readable quote number with organization prefix
          example: "Q-2025-0001"
        customerId:
          type: string
          format: uuid
          description: Customer ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        projectId:
          type: string
          format: uuid
          description: Associated project ID
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          minLength: 3
          maxLength: 255
          description: Quote title
          example: "Website Development Project"
        description:
          type: string
          maxLength: 2000
          description: Quote description
          example: "Complete website development including design, development, and deployment"
        status:
          $ref: '#/components/schemas/QuoteStatus'
        type:
          $ref: '#/components/schemas/QuoteType'
        validFrom:
          type: string
          format: date-time
          description: Quote validity start date
          example: "2025-01-01T00:00:00Z"
        validUntil:
          type: string
          format: date-time
          description: Quote validity end date
          example: "2025-12-31T23:59:59Z"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: Quote currency
          example: "NZD"
        exchangeRate:
          type: number
          minimum: 0
          description: Exchange rate to base currency
          example: 1.0
        subtotal:
          type: string
          description: Subtotal amount as decimal string
          example: "8000.00"
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          description: Quote-level tax rate
          example: 0.15
        taxAmount:
          type: string
          description: Total tax amount as decimal string
          example: "1200.00"
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Quote-level discount value
          example: 0
        discountAmount:
          type: string
          description: Total discount amount as decimal string
          example: "0.00"
        totalAmount:
          type: string
          description: Total amount as decimal string
          example: "9200.00"
        termsConditions:
          type: string
          maxLength: 5000
          description: Terms and conditions
          example: "Payment terms: 50% upfront, 50% on completion"
        notes:
          type: string
          maxLength: 2000
          description: Customer-facing notes
          example: "Includes 3 months of support after launch"
        internalNotes:
          type: string
          maxLength: 2000
          description: Internal notes (not visible to customer)
          example: "Customer prefers React over Vue"
        createdBy:
          type: string
          format: uuid
          description: User who created the quote
          example: "550e8400-e29b-41d4-a716-446655440003"
        approvedBy:
          type: string
          format: uuid
          description: User who approved the quote
          example: "550e8400-e29b-41d4-a716-446655440004"
        approvedAt:
          type: string
          format: date-time
          description: When the quote was approved
          example: "2025-01-15T10:30:00Z"
        sentAt:
          type: string
          format: date-time
          description: When the quote was sent to customer
          example: "2025-01-16T14:00:00Z"
        acceptedAt:
          type: string
          format: date-time
          description: When the quote was accepted by customer
          example: "2025-01-20T09:15:00Z"
        expiresAt:
          type: string
          format: date-time
          description: When the quote expires
          example: "2025-02-01T23:59:59Z"
        organizationId:
          type: string
          format: uuid
          description: Organization ID for multi-tenancy
          example: "550e8400-e29b-41d4-a716-446655440005"
        metadata:
          type: object
          description: Additional metadata
          example: { "source": "sales_team", "priority": "high" }
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-01-01T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-01-15T10:30:00Z"
        lineItems:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/LineItem'
              - type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    description: Line item ID
                    example: "550e8400-e29b-41d4-a716-446655440006"
                  subtotal:
                    type: string
                    description: Line item subtotal
                    example: "6000.00"
                  taxAmount:
                    type: string
                    description: Line item tax amount
                    example: "900.00"
                  discountAmount:
                    type: string
                    description: Line item discount amount
                    example: "0.00"
                  totalAmount:
                    type: string
                    description: Line item total amount
                    example: "6900.00"
                  createdAt:
                    type: string
                    format: date-time
                    description: Creation timestamp
                    example: "2025-01-01T08:00:00Z"
                  updatedAt:
                    type: string
                    format: date-time
                    description: Last update timestamp
                    example: "2025-01-01T08:00:00Z"
...
    UpdateQuoteRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
          description: Quote title
          example: "Updated Website Development Project"
        description:
          type: string
          maxLength: 2000
          description: Quote description
          example: "Updated scope including mobile app development"
        type:
          $ref: '#/components/schemas/QuoteType'
        validFrom:
          type: string
          format: date-time
          description: Quote validity start date
          example: "2025-01-01T00:00:00Z"
        validUntil:
          type: string
          format: date-time
          description: Quote validity end date
          example: "2025-12-31T23:59:59Z"
        currency:
          type: string
          minLength: 3
          maxLength: 3
          description: Quote currency
          example: "NZD"
        exchangeRate:
          type: number
          minimum: 0
          description: Exchange rate to base currency
          example: 1.0
        taxRate:
          type: number
          minimum: 0
          maximum: 1
          description: Quote-level tax rate
          example: 0.15
        discountType:
          $ref: '#/components/schemas/DiscountType'
        discountValue:
          type: number
          minimum: 0
          description: Quote-level discount value
          example: 0
        termsConditions:
          type: string
          maxLength: 5000
          description: Terms and conditions
          example: "Payment terms: 50% upfront, 50% on completion"
        notes:
          type: string
          maxLength: 2000
          description: Customer-facing notes
          example: "Includes 3 months of support after launch"
        internalNotes:
          type: string
          maxLength: 2000
          description: Internal notes (not visible to customer)
          example: "Customer prefers React over Vue"
        lineItems:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/LineItem'
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "ValidationError"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request data"
        code:
          type: string
          description: Error code for programmatic handling
          example: "INVALID_INPUT"
        details:
          type: object
          description: Additional error details
          example:
            field: "title"
            value: ""
            constraint: "minLength"
paths:
...
  /v1/quotes/{id}:
    get:
      summary: Get quote by ID
      description: |
        Retrieve a specific quote by its ID.
        
        ## Access Control
        - Users can only access quotes within their organization
        - Returns 404 if quote doesn't exist or belongs to different organization
      tags:
        - Quotes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Quote ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Quote details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '404':
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Tenant access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update quote
      description: |
        Update an existing quote.
        
        ## Features
        - Partial updates supported (only send fields to change)
        - Automatic recalculation of totals when line items change
        - Audit logging of all changes
        - Status validation (can't update accepted/rejected quotes)
        
        ## Calculations
        When line items are updated, the quote totals are automatically recalculated:
        - Subtotal: Sum of all line item totals
        - Tax: Applied at quote level and line item level
        - Discounts: Can be percentage or fixed amount
        - Total: Subtotal + Tax - Discounts
      tags:
        - Quotes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Quote ID
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuoteRequest'
            examples:
              update_title:
                summary: Update quote title
                value:
                  title: "Updated Website Development Project"
                  description: "Updated scope including mobile app development"
              update_line_items:
                summary: Update line items
                value:
                  lineItems:
                    - lineNumber: 1
                      type: "service"
                      description: "Updated Web Development"
                      quantity: 50
                      unitPrice:
                        amount: "160.00"
                        currency: "NZD"
                      unitCost:
                        amount: "80.00"
                        currency: "NZD"
                      taxRate: 0.15
                      discountType: "percentage"
                      discountValue: 0
                      serviceCategoryId: "550e8400-e29b-41d4-a716-446655440000"
                      metadata:
                        projectPhase: "development"
                        priority: "high"
      responses:
        '200':
          description: Quote updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - quote cannot be updated in current status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Tenant access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...